The issue was definitely with alignment. This will make the unrolling of gebp a fair amount more complicated

+++ Linear/Main.cpp
@@ -30,7 +30,7 @@ int main()
 	};
 
 
-	Matrix<double> bigger = {
+	Matrix<int> bigger = {
 		{ 1,  2,  3,  4,  5 },	
 		{ 6,  7,  8,  9,  10},
 		{ 11, 12, 13, 14, 15},
@@ -62,7 +62,7 @@ int main()
 
 	//Matrix<int> large(100, 100);
 	s.start();
-	Matrix<double> vv = bigger * bigger;
+	Matrix<int> vv = bigger * bigger;
 	s.printCurrent();
 
 	std::cout << vv << "\n\n";
